refactor(core): move empty_bytearray to trezor.utils
