+++ core/src/apps/bitcoin/sign_tx/bitcoin.py
@@ -6,7 +6,7 @@ from trezor.messages import InputScriptType, OutputScriptType
 from trezor.messages.TxRequest import TxRequest
 from trezor.messages.TxRequestDetailsType import TxRequestDetailsType
 from trezor.messages.TxRequestSerializedType import TxRequestSerializedType
-from trezor.utils import HashWriter, ensure
+from trezor.utils import HashWriter, empty_bytearray, ensure
 
 from apps.common.writers import write_bitcoin_varint
 
@@ -93,7 +93,7 @@ class Bitcoin:
         self.external: set[int] = set()
 
         # transaction and signature serialization
-        self.serialized_tx = writers.empty_bytearray(_MAX_SERIALIZED_CHUNK_SIZE)
+        self.serialized_tx = empty_bytearray(_MAX_SERIALIZED_CHUNK_SIZE)
         self.tx_req = TxRequest()
         self.tx_req.details = TxRequestDetailsType()
         self.tx_req.serialized = TxRequestSerializedType()
