+++ core/src/apps/common/cbor.py
@@ -7,7 +7,7 @@ from micropython import const
 
 from trezor import log, utils
 
-from . import readers, writers
+from . import readers
 
 if False:
     from typing import Any, Union, Iterator, Tuple
@@ -250,7 +250,7 @@ def encode_chunked(value: Value, max_chunk_size: int) -> Iterator[bytes]:
 
     chunks = encode_streamed(value)
 
-    chunk_buffer = writers.empty_bytearray(max_chunk_size)
+    chunk_buffer = utils.empty_bytearray(max_chunk_size)
     try:
         current_chunk_view = utils.BufferReader(next(chunks))
         while True:
